Clean up usage of sorting/limit/reverse an extract to pure functions.
